feat: Update Propiedades component to assign authorization to property
This commit updates the Propiedades component to include the functionality to assign authorization to a property. It makes an API call to assign the authorization and then updates the list of authorization documents accordingly. If the assignment is successful, a success toast message is displayed. If there is an error, an error toast message is displayed.