feat: Add ConfirmModal for deleting users in DataTable component
This commit adds a ConfirmModal component to the DataTable component in order to confirm the deletion of a user. When the delete button is clicked, the ConfirmModal is displayed, and if the user confirms the deletion, the handleDelete function is called to delete the user from the list. This improves the user experience by providing a confirmation step before deleting a user.